// app/core/models/Consultantes.ts
export class Consultantes {
  id: number;
  nombre: string;
  apellidos: string;
  email: string;
  idEmpresa: number;
  primerIngreso: boolean;
  tipoDocumento: string;
  numDocumento: number;
  fechaNacimiento: Date;
  estadoCivil: string;
  ciudad: string;
  direccion: string;
  telefono: number;
  imagen: string;
  genero: string;
  idInvitacion: number;
  tpCerebro: string;
  dVivir: string;
}